Move network to another file, make as singletone.
